+++ src/interfaces/Events.ts
@@ -1,6 +1,6 @@
 /// src/interfaces/Events.ts
-import { EventHandler } from "../core/events/EventDispatcherImpl.js"; // Wird gleich angepasst
-import { EventType } from "../enums/EventType.js";
+import { EventHandler } from "../core/index.js"; // Wird gleich angepasst
+import { EventType } from "../enums/index.js";
 
 export interface Events {
   addEventListener(type: string | EventType, listener: EventHandler): void;