+++ src/renderers/AbstractRenderer.ts
@@ -13,7 +13,6 @@ import { LightType, RendererType } from "../enums/index.js";
 import { Object3D } from "../core/Object3D.js";
 import { Scene } from "../core/Scene.js";
 import { Vector3D } from "../math/Vector3D.js";
-
 export abstract class AbstractRenderer implements Renderer {
   public abstract readonly type: RendererType;
   protected _clearColor: Color = new Color(0, 0, 0, 1);