+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [0.11.11] - 2026-03-25
+
 - Apply AGENTS.md
 - Start with nice 2D features
 - Reorganize code